python/requirements.txt
black
isort

pytest
PyHamcrest

nanobind

torch
